调整用户内容解密
